requirements.txt
streamlit
torch
torchvision
tensorflow
Pillow
opencv-python-headless
numpy
pandas
matplotlib
scikit-learn
